requirements.txt
amqp==5.3.1
asgiref==3.11.0
attrs==25.3.0
autobahn==25.11.1
Automat==25.4.16
billiard==4.2.1
cbor2==5.7.1
celery==5.5.3
certifi==2022.12.7
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.0.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
crispy-bootstrap5==2024.10
cron_descriptor==2.0.6
cryptography==46.0.3
daphne==4.2.1
defusedxml==0.7.1
diff-match-patch==20241021
dj-rest-auth==7.0.1
Django==5.1.2
django-allauth==65.0.2
django-celery-beat==2.8.1
django-cors-headers==4.8.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-filter==25.1
django-import-export==4.3.9
django-phonenumber-field==8.1.0
django-phonenumbers==1.0.1
django-timezone-field==7.1
django-unfold==0.65.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.9.1
drf-yasg==1.21.10
gunicorn==23.0.0
h11==0.16.0
hyperlink==21.0.0
idna==3.4
Incremental==24.11.0
inflection==0.5.1
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
loguru==0.7.3
MarkupSafe==3.0.2
msgpack==1.1.2
oauthlib==3.2.2
packaging==23.1
phonenumbers==9.0.13
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg==3.2.3
psycopg-binary==3.2.3
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==25.3.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
requests==2.28.2
requests-oauthlib==1.3.1
rpds-py==0.27.1
service-identity==24.2.0
six==1.17.0
sqlparse==0.4.3
tablib==3.8.0
Twisted==25.5.0
txaio==25.12.1
typing_extensions==4.9.0
tzdata==2025.2
ujson==5.11.0
uritemplate==4.2.0
urllib3==1.26.14
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
zope.interface==8.1.1
